# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactRegistryCreate": {
        "description": "建制品库入参。docker 库配 url;credential 收明文由 API 换 vault id(不落明文)。",
        "properties": {
          "credential": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential"
          },
          "description": {
            "default": "",
            "maxLength": 512,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ArtifactRegistryType"
          },
          "url": {
            "default": "",
            "maxLength": 512,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "title": "ArtifactRegistryCreate",
        "type": "object"
      },
      "ArtifactRegistryType": {
        "description": "制品库类型:docker 镜像仓 / 通用文件仓(tar 包等)。按需再扩 pypi/npm。",
        "enum": [
          "docker",
          "generic"
        ],
        "title": "ArtifactRegistryType",
        "type": "string"
      },
      "BuildNodeCreate": {
        "description": "注册本地或 SSH 构建节点；外部节点需提供 host 与凭证引用。",
        "properties": {
          "host": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "labels": {
            "additionalProperties": true,
            "title": "Labels",
            "type": "object"
          },
          "max_concurrent": {
            "default": 1,
            "maximum": 64.0,
            "minimum": 1.0,
            "title": "Max Concurrent",
            "type": "integer"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "server_id": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Id"
          },
          "ssh_credential_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssh Credential Id"
          }
        },
        "required": [
          "name"
        ],
        "title": "BuildNodeCreate",
        "type": "object"
      },
      "BuildRequestBody": {
        "description": "UI 触发构建入参:可覆盖服务 build_config 的默认 ref/version;缺省则用配置默认。",
        "properties": {
          "git_ref": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Ref"
          },
          "version": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "title": "BuildRequestBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "一脉 Axon 控制面",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/artifact-registries": {
      "get": {
        "operationId": "list_artifact_registries_api_artifact_registries_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Artifact Registries Api Artifact Registries Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Artifact Registries",
        "tags": [
          "builds"
        ]
      },
      "post": {
        "operationId": "create_artifact_registry_api_artifact_registries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactRegistryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Artifact Registry Api Artifact Registries Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Artifact Registry",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/artifact-registries/{registry_id}": {
      "delete": {
        "operationId": "delete_artifact_registry_api_artifact_registries__registry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "registry_id",
            "required": true,
            "schema": {
              "title": "Registry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Artifact Registry Api Artifact Registries  Registry Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Artifact Registry",
        "tags": [
          "builds"
        ]
      }
    },
...
    "/api/build-nodes": {
      "get": {
        "operationId": "list_build_nodes_api_build_nodes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Build Nodes Api Build Nodes Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Build Nodes",
        "tags": [
          "builds"
        ]
      },
      "post": {
        "operationId": "create_build_node_api_build_nodes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildNodeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Build Node Api Build Nodes Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Build Node",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/build-nodes/{node_id}": {
      "delete": {
        "operationId": "delete_build_node_api_build_nodes__node_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Build Node Api Build Nodes  Node Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Build Node",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/build-nodes/{node_id}/heartbeat": {
      "post": {
        "operationId": "heartbeat_build_node_api_build_nodes__node_id__heartbeat_post",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Heartbeat Build Node Api Build Nodes  Node Id  Heartbeat Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Heartbeat Build Node",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/builds/{build_id}": {
      "get": {
        "operationId": "get_build_api_builds__build_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "title": "Build Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Build Api Builds  Build Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Build",
        "tags": [
          "builds"
        ]
      }
    },
...
    "/api/services/{service_id}/artifacts": {
      "get": {
        "operationId": "list_service_artifacts_api_services__service_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Service Artifacts Api Services  Service Id  Artifacts Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Service Artifacts",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/services/{service_id}/build": {
      "post": {
        "description": "触发一次控制面本地构建(方案 A)。未配 build_config 的服务拒绝(501)。",
        "operationId": "trigger_build_api_services__service_id__build_post",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Trigger Build Api Services  Service Id  Build Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trigger Build",
        "tags": [
          "builds"
        ]
      }
    },
    "/api/services/{service_id}/builds": {
      "get": {
        "operationId": "list_service_builds_api_services__service_id__builds_get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Service Builds Api Services  Service Id  Builds Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Service Builds",
        "tags": [
          "builds"
        ]
      }
    },
...
  }
}